Mega Patch Notes 0.2.1
Atualização by Gabriel e Kelvin :D

+++ src/model/beans/Viagem.java
@@ -12,7 +12,7 @@ public class Viagem {
     private int qtVagas;
     private double valor;
     private int ativo;
-    private int despesa;
+    private Double despesa;
     private int pessoas;
 
     public String getVagas() {
@@ -95,11 +95,11 @@ public class Viagem {
         this.ativo = ativo;
     }
 
-    public int getDespesa() {
+    public Double getDespesa() {
         return despesa;
     }
 
-    public void setDespesa(int despesa) {
+    public void setDespesa(Double despesa) {
         this.despesa = despesa;
     }
